+++ examples/c/10-metadata.c
@@ -74,7 +74,6 @@ static gboolean meta_struc_print (GQuark field_id,
   says there is data for us. user_data has to be defined when calling g_signal_connect.
  */
 static GstFlowReturn callback (GstElement* sink, void* user_data)
-
 {
     printf("new sample\n");
 
@@ -129,6 +128,15 @@ static GstFlowReturn callback (GstElement* sink, void* user_data)
 
 int main (int argc, char *argv[])
 {
+    /* this line sets the gstreamer default logging level
+       it can be removed in normal applications
+       gstreamer logging can contain verry useful information
+       when debugging your application
+       # see https://gstreamer.freedesktop.org/documentation/tutorials/basic/debugging-tools.html
+       for further details
+    */
+    gst_debug_set_default_threshold(GST_LEVEL_WARNING);
+
     gst_init(&argc, &argv);
 
     const char* serial = NULL; // set this if you do not want the first found device